requirements.txt
aniso8601==10.0.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
